Initialize the branch
